Die ersten Änderungen am MessageBoard und Notifiable Interfaces wurden umgesetzt. Jeder User hat nun einen UserStatus, welcher in einer Map gespeichert wird. Desweiteren werden die NotifyCallbacks in einem ThreadPool ausgeführt. Dadurch blockiert ein nicht reagierender Callback nicht alle folgenden Callbacks mehr. Wenn ein Callback auf einen Client fehlschlägt, wird der entsprechende Client von der Map entfernt und sein Status wird auf OFFLINE gesetzt. Dann kann er sich neue Nachrichten nur noch durch Polling laden. Außerdem können sich Clienten nun vom Server durch Aufrufen der logout Methode abmelden.

+++ MessageBoard/Server/src/main/java/de/htwsaar/wirth/server/MessageBoardImpl.java
@@ -6,12 +6,18 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
+import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 import de.htwsaar.wirth.remote.MessageBoard;
 import de.htwsaar.wirth.remote.Notifiable;
 import de.htwsaar.wirth.remote.ParentServer;
 import de.htwsaar.wirth.remote.exceptions.MessageNotExistsException;
+import de.htwsaar.wirth.remote.exceptions.NoPermissionException;
+import de.htwsaar.wirth.remote.exceptions.UserNotExistsException;
 import de.htwsaar.wirth.remote.model.MessageImpl;
 import de.htwsaar.wirth.remote.model.Status;
 import de.htwsaar.wirth.remote.model.UserImpl;
@@ -57,10 +63,17 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	/**
 	 * a list of clients
 	 */
-	private List<Notifiable> clientList;
+	private Map<String, Notifiable> clientNotifyMap;
 	
-	// TODO: mein Vorschlag zur Überwachung des online-status
+	/**
+	 * a map to manage the status of each user
+	 */
 	private Map<String, Status> userStatus;
+
+	/**
+	 * a ThreadPool we use for callbacks to clients
+	 */
+	private ExecutorService threadPool;
 
 	private static final long serialVersionUID = -4613549994529764225L;
 
@@ -68,11 +81,15 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 		this.groupName = groupName;
 		childServerList = Collections.synchronizedList(new ArrayList<Notifiable>());
 		childServerQueueMap = new ConcurrentHashMap<Notifiable, CommandRunner>();
-		clientList = Collections.synchronizedList(new ArrayList<Notifiable>());
 		
+		clientNotifyMap = new ConcurrentHashMap<String, Notifiable>();
 		userStatus = new ConcurrentHashMap<String, Status>();
-		for (User user : Services.getInstance().getUserService().getAll())
-			userStatus.put(user.getUsername(), Status.SHOW_AS_OFFLINE);
+		
+		// set the default userstatus of each user to offline
+		for (User user : Services.getInstance().getUserService().getAll()) {
+			userStatus.put(user.getUsername(), Status.OFFLINE);
+		}
+		threadPool = Executors.newCachedThreadPool();
 	}
 	
 	/**
@@ -111,16 +128,17 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @param handler
 	 */
 	private void notifyClients(ClientNotifyHandler handler) {
-		synchronized (clientList) {
-			for (Notifiable client : clientList) {
-				try {
-					handler.handle(client);
-				} catch (RemoteException e) {
-					// Ignore it, if a client cannot be called back
-					e.printStackTrace();
-				}
+		for (Entry<String,Notifiable> entry : clientNotifyMap.entrySet()) {
+				threadPool.execute(() -> {
+					try {
+						handler.handle(entry.getValue());
+					} catch (RemoteException e) {
+						// if we catch a remoteException the callback for this client doesn't work
+						clientNotifyMap.remove(entry.getKey());
+						changeUserStatusAndNotifyClients(entry.getKey(), Status.OFFLINE);
+					}
+				});
 			}
-		}
 	}
 
 	/**
@@ -136,12 +154,17 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 			}
 		}
 	}
+	
+	private void changeUserStatusAndNotifyClients(String username, Status status) {
+		userStatus.put(username, status);
+		notifyClients(cl -> cl.notifyUserStatus(username, status));
+	}
 
 	/**
 	 *	needToSend decides whether a message has to send to the parent or not.
 	 */
 	private boolean needToSendParent(Message msg){
-		// add condition Database has msg
+		// FIXME: add condition Database has msg
 		// wäre es nicht besser einfach grundsätzlich hochzuschieben und wenn der parent die nachricht nicht hat ignoriert er es
 		return (msg.isPublished() && !isRoot());
 	}
@@ -152,10 +175,6 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	private boolean needToPublish(Message msg){
 		// add condition Database has msg
 		return (!msg.isPublished() && !isRoot());
-	}
-
-	private boolean messageExists(Message msg){
-		return Services.getInstance().getMessageService().existsMessage(msg);
 	}
 	
 	//---------------------------------- MessageBoard Interface ----------------------------------
@@ -168,38 +187,31 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @return
 	 * @throws RemoteException
 	 */
-	// TODO: Methode in login umbenennen
-	public AuthPacket registerClient(LoginPacket login, Notifiable client) throws RemoteException {
-		// Authenticate throws an exception, if the username or password are
-		// wrong
+	public AuthPacket login(LoginPacket login, Notifiable client) throws RemoteException {
+		// Authenticate throws an exception, if the username or password are wrong
 		// this exception can be handled on clientside
-		AuthPacket auth = SessionManager.authenticate(login);		
-		// TODO: müsste man hier nicht noch allen Client-Childs bescheid sagen, dass ein neuer User da ist ?
-		// also etwa:		
-//		for (Notifiable c : clientList) {
-//			c.notifyUserStatus(login.getUsername(), Status.ONLINE);
-//		}
-		// und dann natürlich auch beim logout wieder weg s.u.
-		clientList.add(client);
-		userStatus. put(login.getUsername(), Status.ONLINE);
+		AuthPacket auth = SessionManager.authenticate(login);
+		
+		changeUserStatusAndNotifyClients(login.getUsername(), Status.ONLINE);
+		
+		// add the Notifiable of the client to the clientNotifyMap
+		clientNotifyMap.put(login.getUsername(), client);
 		return auth;
 	}
 	
-//	// TODO: logout
-//	public void logout(AuthPacket auth, Notifiable client) {
-//		clientList.remove(client);
-//		userStatus.put(auth.getUsername(), Status.SHOW_AS_OFFLINE);
-//		onlineUsers.remove(login.getUsername());
-//		for (Notifiable c : clientList) {
-//			c.notifyUserStatus(auth.getUsername(), Status.SHOW_AS_OFFLINE);
-//		}
-//	}
+	public void logout(AuthPacket auth) throws RemoteException {
+		SessionManager.verifyAuthPacket(auth);
+		// remove the user from the clientNotifyMap and set his status to offline
+		clientNotifyMap.remove(auth.getUsername());
+		SessionManager.logout(auth.getUsername());
+		
+		changeUserStatusAndNotifyClients(auth.getUsername(), Status.OFFLINE);
+	}
 	
 	public void changeUserStatus(AuthPacket auth, Status status) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
-		for (Notifiable c : clientList) {
-			c.notifyUserStatus(auth.getUsername(), status);
-		}		
+		SessionManager.verifyAuthPacket(auth);
+		String username = auth.getUsername();
+		changeUserStatusAndNotifyClients(username, status);
 	}
 		
 	/**
@@ -208,23 +220,38 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @param newPassword
 	 */
 	public void addUser(AuthPacket auth, String newUsername, String newPassword) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
-		SessionManager.isGroupLeader(auth);
+		SessionManager.verifyAuthPacket(auth);
+		if (SessionManager.isGroupLeader(auth)) {
+			throw new NoPermissionException("The user is not a group-leader");
+	    }
 		User user = new UserImpl(newUsername, "", "", newPassword, false);
 		Services.getInstance().getUserService().saveUser(user);
-		userStatus.put(newUsername, Status.SHOW_AS_OFFLINE);
+		
+		// notify the clients of the new offline user
+		changeUserStatusAndNotifyClients(newUsername, Status.OFFLINE);
 	}
 	
 	public void deleteUser(AuthPacket auth, String username) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
-		SessionManager.isGroupLeader(auth);
+		SessionManager.verifyAuthPacket(auth);
+		if (SessionManager.isGroupLeader(auth)) {
+			throw new NoPermissionException("The user is not a group-leader");
+		}
+		// check if the user exists
+		User user = Services.getInstance().getUserService().getUser(username);
+		
+		// TODO: die Nutzer müssen über einen gelöschten UserStatus benachrichtigt werden.
+		// Dies ist mit unserem derzeitigen Notifiable Interface nicht möglich :(
+		
 //		if (userStatus.get(username) == Status.ONLINE)
 //			throw UserIsOnlineException("The User is still online");
-		// TODO: alternativ auch irgendwie den User rausschmeißen im moment schwer machbar vlt statt List<Notifiable> clients lieber Map<String, Notifiable> mit username als key 
-		User user = Services.getInstance().getUserService().getUser(username);
-//		if ( user == null )
-//			throw UserNotExistsException("This User does not exist on this server");
+		
+		if ( user == null ) {
+			throw new UserNotExistsException("This User does not exist on this server");
+		}
 		Services.getInstance().getUserService().deleteUser(user);
+		
+		SessionManager.logout(username);
+		clientNotifyMap.remove(username);
 		userStatus.remove(username);
 			
 	}
@@ -239,11 +266,13 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @param msg
 	 * @throws RemoteException
 	 */
-	// TODO: void publish(AuthPacket auth, UUID id) throws RemoteException;
-	public void publish(AuthPacket auth, Message msg) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
-		SessionManager.isGroupLeader(auth);
-		if(!messageExists(msg)) {
+	public void publish(AuthPacket auth, UUID id) throws RemoteException {
+		SessionManager.verifyAuthPacket(auth);
+		if (SessionManager.isGroupLeader(auth)) {
+			throw new NoPermissionException("The user is not a group-leader");
+		}
+		Message msg = Services.getInstance().getMessageService().getMessage(id);
+		if ( msg == null ) {
 			throw new MessageNotExistsException("The message doesn't exists on this server");
 		}
 		if(needToPublish(msg)) {
@@ -264,7 +293,7 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @throws RemoteException
 	 */
 	public void newMessage(AuthPacket auth, String msg) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
+		SessionManager.verifyAuthPacket(auth);
 		Message message = new MessageImpl(msg, auth.getUsername(), groupName);
 		notifyNew(message);
 	}
@@ -277,28 +306,25 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * the User a username and a valid authentication token is required.
 	 *
 	 * @param auth AuthPack given by the login
-	 * @param msg
+	 * @param msgTxt
 	 * @throws RemoteException
 	 */
-	// TODO: void editMessage(AuthPacket auth, String msg, UUID id) throws RemoteException;
-	public void editMessage(AuthPacket auth, Message msg) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
-		SessionManager.isAuthor(auth, msg);			
+	public void editMessage(AuthPacket auth, String msgTxt, UUID id) throws RemoteException {
+		SessionManager.verifyAuthPacket(auth);
+		Message oldMsg;
 		synchronized (Message.class) {
-			// TODO: check, if the given message is equal to the msg in the database, except for the modified msgtxt.
-//			Message oldMsg = Services.getInstance().getMessageService().getMessage(id);
-//			if ( oldMsg == null)
-//				throw new MessageNotExistsException("The message doesn't exists on this server");
-						
-			if(!messageExists(msg)) {
+			oldMsg = Services.getInstance().getMessageService().getMessage(id);
+			if ( oldMsg == null)
 				throw new MessageNotExistsException("The message doesn't exists on this server");
+			if (SessionManager.isAuthor(auth, oldMsg)) {
+	            throw new NoPermissionException("The user is not the author"); 
 			}
-//			if ( !oldMsg.getMessage().equals(msg.getMessage()) )
-			//notifyEdit(msg, id);
-			notifyEdit(msg);
+
+			if ( !oldMsg.getMessage().equals(msgTxt) )
+				notifyEdit(oldMsg);
 		}
-		if(needToSendParent(msg)) {
-			Command cmd = CommandBuilder.buildParentCommand(parent, msg, ParentCmd.EDIT);
+		if(needToSendParent(oldMsg)) {
+			Command cmd = CommandBuilder.buildParentCommand(parent, oldMsg, ParentCmd.EDIT);
 			parentQueue.addCommand(cmd);
 		}
 	}	
@@ -313,20 +339,21 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @param msg
 	 * @throws RemoteException
 	 */
-	// TODO: void deleteMessage(AuthPacket auth, UUID id) throws RemoteException;
-	public void deleteMessage(AuthPacket auth, Message msg) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
-		SessionManager.isGroupLeaderOrAuthor(auth, msg);
+	public void deleteMessage(AuthPacket auth, UUID id) throws RemoteException {
+		SessionManager.verifyAuthPacket(auth);
+		Message msgToDelete;
 		synchronized (Message.class) {
-//			if ( Services.getInstance().getMessageService().getMessage(id) == null )
-			if(!messageExists(msg)) {
+			msgToDelete = Services.getInstance().getMessageService().getMessage(id);
+			if (!SessionManager.isAuthor(auth, msgToDelete) && !SessionManager.isGroupLeader(auth)) {
+				throw new NoPermissionException("The user is neither the author nor a groupleader");
+			}
+			if ( Services.getInstance().getMessageService().getMessage(id) == null ) {
 				throw new MessageNotExistsException("The message doesn't exists on this server");
 			}
-			// notifyDelete(id);
-			notifyDelete(msg);
+			notifyDelete(msgToDelete);
 		}
-		if(needToSendParent(msg)) {
-			Command cmd = CommandBuilder.buildParentCommand(parent, msg, ParentCmd.DELETE);
+		if(needToSendParent(msgToDelete)) {
+			Command cmd = CommandBuilder.buildParentCommand(parent, msgToDelete, ParentCmd.DELETE);
 			parentQueue.addCommand(cmd);
 		}
 	}
@@ -341,13 +368,12 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @throws RemoteException
 	 */
 	public List<Message> getMessages(AuthPacket auth) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
+		SessionManager.verifyAuthPacket(auth);
 		return getMessages();
 	}
 		
 	public Map<String, Status> getUserStatus(AuthPacket auth) throws RemoteException {
-		SessionManager.isAuthenticatedByToken(auth);
-		// TODO Ist Status serialisierbar ? und Map ?
+		SessionManager.verifyAuthPacket(auth);
 		return userStatus;
 	}
 	
@@ -370,7 +396,7 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 		queueCommandForAllChildServer(CommandBuilder.buildChildCommand(msg, ChildCmd.NEW));
 
 		// Notify each client
-		notifyClients((cl) -> cl.notifyNew(msg));
+		notifyClients(cl -> cl.notifyNew(msg));
 	}
 
 	/**
@@ -381,20 +407,18 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 *
 	 * @param msg message to be edited
 	 * @throws RemoteException
-	 */
-	// TODO: void notifyEdit(String msg, UUID id) throws RemoteException;		
+	 */		
 	public void notifyEdit(Message msg) throws RemoteException {
-//		Message message = Services.getInstance().getMessageService().getMessage(id);
-//		if ( message != null )
-		if(messageExists(msg)) {
-//			message.changeMessage(msg);
+		Message message = Services.getInstance().getMessageService().getMessage(msg.getID());
+		if ( message != null ) {
+			message.changeMessage(msg.getMessage());
 			Services.getInstance().getMessageService().saveMessage(msg/*message*/);
 
 			// Add a EditMessageCommand to each CommandRunner
 			queueCommandForAllChildServer(CommandBuilder.buildChildCommand(msg, ChildCmd.EDIT));
 
 			// Notify each client
-			notifyClients((cl) -> cl.notifyEdit(msg));
+			notifyClients(cl -> cl.notifyEdit(msg));
 		}
 	}
 
@@ -407,21 +431,16 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @param msg message to be delete
 	 * @throws RemoteException
 	 */
-	// TODO: void notifyDelete(UUID id) throws RemoteException;
 	public void notifyDelete(Message msg) throws RemoteException {
-		if(messageExists(msg)) {
-//		if ( Services.getInstance().getMessageService().getMessage(id) != null )
-//			Services.getInstance().getMessageService().deleteMessage(id);
+		if ( Services.getInstance().getMessageService().getMessage(msg.getID()) != null ) {
 			Services.getInstance().getMessageService().deleteMessage(msg);
-
-			// only execute the following, if the delete was successful
 
 			// Add a DeleteMessageCommand to each CommandRunner
 			queueCommandForAllChildServer(CommandBuilder.buildChildCommand(msg, ChildCmd.DELETE));
 
 			// Notify each client
-			notifyClients((cl) -> cl.notifyDelete(msg));
-			}
+			notifyClients(cl -> cl.notifyDelete(msg));
+		}
 	}
 	
 	//---------------------------------- ParentServer Interface ----------------------------------
@@ -462,9 +481,7 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @param msg
 	 * @throws RemoteException
 	 */
-	// TODO: void notifyServerEdit(String msg, UUID id) throws RemoteException
 	public void notifyServerEdit(Message msg) throws RemoteException {
-		// notifyEdit(msg, id);
 		notifyEdit(msg);
 		if(needToSendParent(msg)) {
 			Command cmd = CommandBuilder.buildParentCommand(parent, msg, ParentCmd.EDIT);
@@ -481,9 +498,7 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 	 * @param msg
 	 * @throws RemoteException
 	 */
-	// TODO:	void notifyServerDelete(UUID id) throws RemoteException
 	public void notifyServerDelete(Message msg) throws RemoteException {
-		// notifyDelete(id);
 		notifyDelete(msg);		
 		if(needToSendParent(msg)) {
 			Command cmd = CommandBuilder.buildParentCommand(parent, msg, ParentCmd.DELETE);
@@ -494,6 +509,8 @@ public class MessageBoardImpl extends UnicastRemoteObject implements Notifiable,
 
 	public List<Message> getMessages() throws RemoteException {
 		// hier evtl noch ein limit fuer nachrichten
+		// TODO: wir sollten jeweils methoden mit limitierung und filterung nach gruppe erstellen
+		// entsprechende methoden sollten die jeweiligen Datenbankcalls nutzen
 		return Services.getInstance().getMessageService().getAll();
 	}
